requirements.txt
swig
gymnasium==0.29.1
lbforaging==2.0.0
pyglet==1.5.29
cleanrl==0.4.8
numpy==1.24.1
tyro==0.8.6
tqdm==4.66.5
swig==4.2.1
gymnasium[box2d]
torch-tb-profiler